requirements.txt
openai
gradio
requests
python-dotenv
beautifulsoup4
tiktoken

